move more buffer-related methods up to ByteArrayBuffer
